+++ sdk/translation/azure-ai-translation-document/samples/async_samples/sample_synchronous_document_translation_async.py
@@ -23,6 +23,7 @@ from azure.ai.translation.document.aio import SingleDocumentTranslationClient
 from azure.ai.translation.document.models import DocumentTranslateContent
 from azure.ai.translation.document._vendor import FileType
 
+
 async def sample_synchronous_document_translation_async():
     # [START synchronous_document_translation_async]
     endpoint = os.environ["AZURE_DOCUMENT_TRANSLATION_ENDPOINT"]
@@ -39,7 +40,7 @@ async def sample_synchronous_document_translation_async():
 
     # Construct the path to the desired file
     file_path = os.path.join(parent_directory, "tests", "TestData", "test-input.txt")
-    file_name = os.path.basename(file_path) 
+    file_name = os.path.basename(file_path)
     file_type = "text/html"
     with open(file_path, "r") as file:
         file_contents = file.read()
@@ -51,10 +52,11 @@ async def sample_synchronous_document_translation_async():
         response_stream = await client.document_translate(
             body=document_translate_content, target_language=target_languages
         )
-    translated_response = response_stream.decode("utf-8-sig")
+    translated_response = response_stream.decode("utf-8-sig") # type: ignore[attr-defined]
     print(f"Translated response: {translated_response}")
 
     # [END synchronous_document_translation_async]
+
 
 async def main():
     await sample_synchronous_document_translation_async()