+++ sdk/translation/azure-ai-translation-document/CHANGELOG.md
@@ -1,12 +1,15 @@
 # Release History
 
-## 1.0.1 (Unreleased)
+## 2.0.0-beta.1 (2024-05-07)
 
 ### Features Added
+- `documenttranslate` is the method added to submit a single/synchronous document translation request to the Document Translation service.
+- `DocumentTranslateContent` has been added to specify the `documenttranslate` request content.
+- Added `getSupportedFormats` that returns a list of supported documents and glossaries by the Document Translation service.
 
 ### Breaking Changes
-
-### Bugs Fixed
+- Version `v1.0` is not supported
+- Renamed property `FormatType` to `FileFormatType`
 
 ### Other Changes